Extract step state tracking into its own module.
This will be utilized in subsequent commits that further flesh out the
states.

LP: #1477904
