+++ snapcraft/tests/test_commands_clean.py
@@ -23,6 +23,7 @@ from snapcraft.main import main
 from snapcraft import (
     common,
     pluginhandler,
+    states,
     tests,
 )
 
@@ -152,15 +153,15 @@ parts:
         main(['clean', '--step=foo'])
 
         expected_staged_state = {
-            'clean0': pluginhandler.StageState({'clean0'}, set()),
-            'clean1': pluginhandler.StageState({'clean1'}, set()),
-            'clean2': pluginhandler.StageState({'clean2'}, set()),
+            'clean0': states.StageState({'clean0'}, set()),
+            'clean1': states.StageState({'clean1'}, set()),
+            'clean2': states.StageState({'clean2'}, set()),
         }
 
         expected_stripped_state = {
-            'clean0': pluginhandler.StripState({'clean0'}, set()),
-            'clean1': pluginhandler.StripState({'clean1'}, set()),
-            'clean2': pluginhandler.StripState({'clean2'}, set()),
+            'clean0': states.StripState({'clean0'}, set()),
+            'clean1': states.StripState({'clean1'}, set()),
+            'clean2': states.StripState({'clean2'}, set()),
         }
 
         mock_clean.assert_called_with(
